@@ -10,8 +10,8 @@
   "homepage": "./",
   "dependencies": {
     "@apollo/client": "^3.3.21",
-    "@dnd-kit/core": "^3.1.1",
-    "@dnd-kit/sortable": "^4.0.0",
+    "@dnd-kit/core": "^6.0.5",
+    "@dnd-kit/sortable": "^7.0.1",
     "@emotion/css": "^11.1.3",
     "@emotion/react": "^11.7.0",
     "@emotion/styled": "^11.6.0",
@@ -21,43 +21,44 @@
     "@mui/styles": "^5.2.3",
     "@reduxjs/toolkit": "^1.6.0",
     "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
     "ansi-to-react": "^6.1.6",
     "classnames": "^2.3.1",
     "crypto-js": "^4.0.0",
     "diff2html": "^3.4.8",
     "formik": "^2.2.9",
-    "framer-motion": "^4",
-    "graphql": "^15.5.1",
-    "monaco-editor": "^0.26.1",
-    "monaco-editor-webpack-plugin": "^4.1.1",
-    "nanoid": "^3.1.23",
+    "framer-motion": "^7.3.5",
+    "graphql": "^16.6.0",
+    "monaco-editor": "^0.34.0",
+    "monaco-editor-webpack-plugin": "^7.0.1",
+    "nanoid": "^4.0.0",
     "nanoid-dictionary": "^4.3.0",
+    "net": "^1.0.2",
     "notistack": "^2.0.3",
     "prismjs": "^1.24.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-app-rewired": "^2.1.8",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.2.0",
-    "react-monaco-editor": "^0.44.0",
-    "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "react-monaco-editor": "^0.50.1",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.4.0",
+    "react-scripts": "5.0.1",
     "redux": "^4.1.0",
-    "slate": "^0.63.0",
-    "slate-history": "^0.62.0",
-    "slate-react": "^0.65.2",
+    "slate": "^0.82.1",
+    "slate-history": "^0.66.0",
+    "slate-react": "^0.82.2",
     "socket.io-client": "^4.1.3",
     "stompjs": "^2.3.3",
     "strip-ansi": "^7.0.0",
-    "use-resize-observer": "^7.0.0",
-    "web-vitals": "^1.0.1",
+    "use-resize-observer": "^9.0.2",
+    "web-vitals": "^3.0.1",
     "xterm": "^4.13.0",
     "xterm-addon-fit": "^0.5.0"
   },
   "scripts": {
-    "start": "PORT=14322 react-app-rewired start",
+    "start": "PORT=3000 react-app-rewired start",
     "build": "react-app-rewired --max_old_space_size=4096 build",
     "test": "react-app-rewired test",
     "eject": "react-app-rewired eject",
@@ -107,10 +108,10 @@
     ]
   },
   "devDependencies": {
-    "concurrently": "^6.3.0",
+    "concurrently": "^7.4.0",
     "cross-env": "^7.0.3",
-    "electron": "^15.3.1",
-    "electron-builder": "^22.13.1",
+    "electron": "^20.1.3",
+    "electron-builder": "^23.3.3",
     "electron-rebuild": "^3.2.5",
     "electronmon": "^2.0.2",
     "wait-on": "^6.0.0"
