@@ -1,6 +1,6 @@
 import "./App.css";
 
-import { HashRouter as Router, Switch, Route } from "react-router-dom";
+import { HashRouter as Router, Routes, Route } from "react-router-dom";
 
 import Home from "./pages/index";
 import About from "./pages/about";
@@ -50,56 +50,83 @@ function App() {
         <ThemeProvider theme={theme}>
           <AuthProvider>
             <SnackbarProvider maxSnack={5}>
-              <Switch>
-                <Route path="/about">
-                  <NormalLayout>
-                    <About />
-                  </NormalLayout>
-                </Route>
-                <Route path="/docs">
-                  <NormalLayout>
-                    <Docs />
-                  </NormalLayout>
-                </Route>
-                <Route path="/repos">
-                  <NormalLayout>
-                    <Repos />
-                  </NormalLayout>
-                </Route>
-                <Route path="/:username/:reponame">
-                  <Box height="100vh">
-                    <Header />
-                    <Box height="100%" pt="50px">
-                      <Repo />
+              <Routes>
+                <Route
+                  path="/about"
+                  element={
+                    <NormalLayout>
+                      <About />
+                    </NormalLayout>
+                  }
+                />
+                <Route
+                  path="/docs"
+                  element={
+                    <NormalLayout>
+                      <Docs />
+                    </NormalLayout>
+                  }
+                ></Route>
+                <Route
+                  path="/repos"
+                  element={
+                    <NormalLayout>
+                      <Repos />
+                    </NormalLayout>
+                  }
+                ></Route>
+                <Route
+                  path="/repo/:id"
+                  element={
+                    <Box height="100vh">
+                      <Header />
+                      <Box height="100%" pt="50px">
+                        <Repo />
+                      </Box>
                     </Box>
-                  </Box>
-                </Route>
-                <Route path="/test">
-                  <NormalLayout>
-                    <Test />
-                  </NormalLayout>
-                </Route>
-                <Route path="/login">
-                  <NormalLayout>
-                    <Login />
-                  </NormalLayout>
-                </Route>
-                <Route path="/signup">
-                  <NormalLayout>
-                    <Signup />
-                  </NormalLayout>
-                </Route>
-                <Route path="/profile">
-                  <NormalLayout>
-                    <Profile />
-                  </NormalLayout>
-                </Route>
-                <Route path="/">
-                  <NormalLayout>
-                    <Home />
-                  </NormalLayout>
-                </Route>
-              </Switch>
+                  }
+                ></Route>
+                <Route
+                  path="/test"
+                  element={
+                    <NormalLayout>
+                      <Test />
+                    </NormalLayout>
+                  }
+                ></Route>
+                <Route
+                  path="/login"
+                  element={
+                    <NormalLayout>
+                      <Login />
+                    </NormalLayout>
+                  }
+                ></Route>
+                <Route
+                  path="/signup"
+                  element={
+                    <NormalLayout>
+                      <Signup />
+                    </NormalLayout>
+                  }
+                ></Route>
+                <Route
+                  path="/profile"
+                  element={
+                    <NormalLayout>
+                      <Profile />
+                    </NormalLayout>
+                  }
+                ></Route>
+                <Route
+                  path="/"
+                  element={
+                    <NormalLayout>
+                      <Home />
+                    </NormalLayout>
+                  }
+                ></Route>
+              </Routes>
             </SnackbarProvider>
           </AuthProvider>
         </ThemeProvider>
