@@ -50,12 +50,7 @@ import { v4 as uuidv4 } from "uuid";
 import * as Diff2Html from "diff2html";
 import "diff2html/bundles/css/diff2html.min.css";
 
-import {
-  repoSlice,
-  loadPodQueue,
-  remoteUpdatePod,
-  selectIsDirty,
-} from "../../lib/store";
+import { repoSlice, remoteUpdatePod, selectIsDirty } from "../../lib/store";
 import * as wsActions from "../../lib/ws/actions";
 import * as qActions from "../../lib/queue/actions";
 
@@ -807,7 +802,7 @@ export function HoveringMenu({ pod, showMenu, draghandle, children }) {
           </HStack>
           <HStack my={2}>
             <InfoBar pod={pod} />
-            <TypeMenu pod={pod} />
+            {pod.type !== "DECK" && <TypeMenu pod={pod} />}
             <LanguageMenu pod={pod} />
             <Button
               size="small"
