@@ -52,7 +52,10 @@ function Repos() {
   if (error) return <Alert severity="error">{error.message}</Alert>;
   let repos = data.myRepos.slice().reverse();
   return (
-    <Box>
+    <Box sx={{ pt: 4 }}>
+      <Box>
+        Hello, {me?.firstname}! Your ID: {me?.id}
+      </Box>
       <CreateRepoForm />
       <Typography variant="h2" gutterBottom component="div">
         Your repos {repos.length}:
@@ -62,8 +65,8 @@ function Repos() {
           The link:{" "}
           <Link
             component={ReactLink}
-            to={`/${me?.username}/${repo.name}`}
-          >{`/${me?.username}/${repo.name}`}</Link>
+            to={`/repo/${repo.id}`}
+          >{`${repo.name}`}</Link>
           <Button
             size="xs"
             sx={{
@@ -90,7 +93,9 @@ function CreateRepoForm(props) {
   const [createRepo, {}] = useMutation(
     gql`
       mutation CreateRepo($name: String!) {
-        createRepo(name: $name)
+        createRepo(name: $name) {
+          name
+        }
       }
     `,
     {
